pyproject.toml
[tool.poetry]
name = "medpack"
version = "0.1.0"
description = ""
authors = ["kai nagae <muhammadfaraz9736@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
django-environ = "^0.6.0"
python-slugify = "^5.0.2"
argon2-cffi = "^21.1.0"
whitenoise = "^5.3.0"
celery = "^5.1.2"
redis = "^3.5.3"
django-celery-beat = "^2.2.1"
django-model-utils = "^4.1.1"
django-allauth = "^0.45.0"
django-crispy-forms = "^1.12.0"
django-redis = "^5.0.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.8.0"
drf-yasg = "^1.20.0"
psycopg2 = "^2.9.1"
graphene-django = "^2.15.0"
django-graphql-jwt = "0.3.0"
django-graphql-auth = "^0.3.16"
Pillow = "^8.3.2"
django-nested-admin = "^3.3.3"
djoser = "^2.1.0"

[tool.poetry.dev-dependencies]
django-werkzeug = "^1.0.0"
ipdb = "^0.13.9"
watchgod = "^0.7"
mypy = "^0.910"
django-stubs = "^1.9.0"
pytest = "^6.2.5"
pytest-sugar = "^0.9.4"
inmanta-sphinx = "^1.4.3"
sphinx-autobuild = "^2021.3.14"
flake8 = "^3.9.2"
flake8-isort = "^4.0.0"
coverage = "^5.5"
black = "^21.8b0"
pylint-django = "^2.4.4"
pylint-celery = "^0.3"
pre-commit = "^2.15.0"
factory-boy = "^3.2.0"
django-debug-toolbar = "^3.2.2"
django-extensions = "^3.1.3"
django-coverage-plugin = "^2.0.0"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
